Test: Add staged feature
This will test the worst case scenario for the staged mode, line number
changing above and below.

Don't force MacOS vim path

Make it easier to override which VIM executable is used, while keeping
the current default.

Provide a way to see staged changes

Improve the signals handling for different modes

Don't clobber the existing signals, keep separate sets of signals and
choose at runtime when to use each.

Initial support for hunks on staged mode

This includes initial, alpha support to "revert" staged changes. What is
really does is reset the hunk.

There are still problems. Since the staged mode hunk list is
non-existant/wrong, the positions are not corrected, so there's an
offset. If you correct that offset manually (relaxed current hunk
checks), the patch is applied correctly and the chunks are reset.

The priority now should be generating the correct hunks for the staged
mode.

Fix hunk generation for the new functions

Add missing staged signs

Add missing endif

Actually update the signals in the buffer